# tests/contracts/fixtures/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ComplianceIQ Core API",
    "description": "\nThe authoritative Core Platform & Data Service for ComplianceIQ.\n\n**Core owns the truth.** Findings, resources, scans and compliance scores\nare produced and owned here. The AI Service reasons *over* this data via\nthis API and never becomes its owner; the dashboard presents it and must\nkeep working when the AI Service is unavailable.\n\n### Authentication\nEvery `/api/v1` endpoint requires `Authorization: Bearer <JWT>` (RS256).\nThe **tenant is taken from the verified token** \u2014 there is no `tenant_id`\nparameter anywhere in this API, and one cannot be supplied.\n\n### Correlation\nSend `X-Correlation-ID` and it is preserved and echoed; omit it and one\nis generated. It appears on every response, including errors.\n\n### Errors\nEvery error \u2014 including validation and unexpected failures \u2014 returns the\nsame envelope: `{\"error\": {\"code\", \"message\", \"correlation_id\", \"details\"}}`.\nBranch on `code`; it is stable. `message` is for humans.\n\n### Versioning\nBreaking changes require a new prefix (`/api/v2`). Additive fields may\nappear in `v1` responses, so clients must ignore unknown fields.\n",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/scores": {
      "get": {
        "tags": [
          "scores"
        ],
        "summary": "List compliance scores",
        "description": "One page of scores for the authenticated tenant, newest first.\n\n**`score` may be `null`.** That means nothing determinate was evaluated in that scope. Render it as 'no data' \u2014 coercing it to 0 or 100 states something false about the tenant's posture.\n\nPair it with `coverage`: a 100% score computed over 4 of 900 checks is an absent posture, not a good one.",
        "operationId": "list_scores_api_v1_scores_get",
        "parameters": [
          {
            "name": "scope",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ScoreScope"
                },
                {
                  "type": "null"
                }
              ],
              "description": "tenant | framework | domain | scan",
              "title": "Scope"
            },
            "description": "tenant | framework | domain | scan"
          },
          {
            "name": "scope_value",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "The framework id / domain name / scan key. Requires scope.",
              "title": "Scope Value"
            },
            "description": "The framework id / domain name / scan key. Requires scope."
          },
          {
            "name": "scan_key",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Scan Key"
            }
          },
          {
            "name": "computed_after",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Computed After"
            }
          },
          {
            "name": "computed_before",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Computed Before"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Page size. Default 50, maximum 100.",
              "default": 50,
              "title": "Limit"
            },
            "description": "Page size. Default 50, maximum 100."
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Zero-based offset.",
              "default": 0,
              "title": "Offset"
            },
            "description": "Zero-based offset."
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponse_ComplianceScoreResource_"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Lacking the 'reader' role.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Invalid filter or pagination bound.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/v1/scores/current": {
      "get": {
        "tags": [
          "scores"
        ],
        "summary": "Get the current score for one scope",
        "description": "The dashboard's headline number. Defaults to the whole-tenant score.",
        "operationId": "current_score_api_v1_scores_current_get",
        "parameters": [
          {
            "name": "scope",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/ScoreScope",
              "default": "tenant"
            }
          },
          {
            "name": "scope_value",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Framework id / domain name / scan key. Omit for scope=tenant.",
              "title": "Scope Value"
            },
            "description": "Framework id / domain name / scan key. Omit for scope=tenant."
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComplianceScoreResource"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Lacking the 'reader' role.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Invalid filter or pagination bound.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "No score has been computed for this scope yet (no scan has completed).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ComplianceScoreResource": {
        "properties": {
          "tenant_id": {
            "type": "string",
            "title": "Tenant Id"
          },
          "scope": {
            "type": "string",
            "enum": [
              "tenant",
              "framework",
              "domain",
              "scan"
            ],
            "title": "Scope"
          },
          "scope_value": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scope Value"
          },
          "score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Score",
            "description": "Percent of DETERMINATE checks passed, or null."
          },
          "coverage": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Coverage",
            "description": "Percent of checks that could be determined, or null."
          },
          "counts": {
            "$ref": "#/components/schemas/ScoreCountsResource"
          },
          "computed_at": {
            "type": "string",
            "format": "date-time",
            "title": "Computed At"
          },
          "scan_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scan Key"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "tenant_id",
          "scope",
          "scope_value",
          "score",
          "coverage",
          "counts",
          "computed_at"
        ],
        "title": "ComplianceScoreResource",
        "description": "A compliance score.\n\n``score`` is nullable and that is load-bearing: ``null`` means\nnothing determinate was evaluated. Clients must render it as \"no\ndata\", never coerce it to 0 or 100 \u2014 both are false statements about\nthe tenant's posture.\n\n``coverage`` is the honesty companion: a 100% score over 4 of 900\nchecks is not a good posture, and only ``coverage`` reveals that."
      },
      "ErrorDetail": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code",
            "description": "Stable machine-readable error code. Branch on this."
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Human-readable explanation. May be reworded; do not parse."
          },
          "correlation_id": {
            "type": "string",
            "title": "Correlation Id",
            "description": "Echoes X-Correlation-ID; quote it in bug reports."
          },
          "details": {
            "additionalProperties": true,
            "type": "object",
            "title": "Details"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "code",
          "message",
          "correlation_id"
        ],
        "title": "ErrorDetail"
      },
      "ErrorEnvelope": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope",
        "description": "The response body of EVERY error this API returns."
      },
...
      "PageResponse_ComplianceScoreResource_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ComplianceScoreResource"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          },
          "has_more": {
            "type": "boolean",
            "title": "Has More"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset",
          "has_more"
        ],
        "title": "PageResponse[ComplianceScoreResource]"
      },
...
      "ScoreCountsResource": {
        "properties": {
          "passed": {
            "type": "integer",
            "title": "Passed"
          },
          "failed": {
            "type": "integer",
            "title": "Failed"
          },
          "indeterminate": {
            "type": "integer",
            "title": "Indeterminate"
          },
          "critical": {
            "type": "integer",
            "title": "Critical"
          },
          "high": {
            "type": "integer",
            "title": "High"
          },
          "medium": {
            "type": "integer",
            "title": "Medium"
          },
          "low": {
            "type": "integer",
            "title": "Low"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "passed",
          "failed",
          "indeterminate",
          "critical",
          "high",
          "medium",
          "low"
        ],
        "title": "ScoreCountsResource"
      },
      "ScoreScope": {
        "type": "string",
        "enum": [
          "tenant",
          "framework",
          "domain",
          "scan"
        ],
        "title": "ScoreScope",
        "description": "What a score is measuring.\n\nA closed vocabulary because each value implies a different\n``scope_value`` semantic, and an open-ended scope string would make\nthe stored rows impossible to interpret without guessing."
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "RS256 JWT issued by complianceiq-core. Required claims: sub, tenant_id, roles, iss=complianceiq-core, aud=complianceiq, exp. Verify against /.well-known/jwks.json."
      }
    }
  }
}